Write a YAML file

Write the YAML representation of an R object to a file

Usage

write_yaml(x, file, fileEncoding = "UTF-8", ...)

Arguments

x

The object to be converted.

file

Either a character string naming a file or a connection open for writing.

fileEncoding

Character string: if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded as they are written. See file().

...

Arguments to as.yaml().

Details

If file is a non-open connection, an attempt is made to open it and then close it after use.

This function is a convenient wrapper around as.yaml().

See also

Author

Jeremy Stephens jeremy.f.stephens@vumc.org

Examples

if (FALSE) { # \dontrun{
  # writing to a file connection
  filename <- tempfile()
  con <- file(filename, "w")
  write_yaml(data.frame(a=1:10, b=letters[1:10], c=11:20), con)
  close(con)

  # using a filename to specify output file
  write_yaml(data.frame(a=1:10, b=letters[1:10], c=11:20), filename)
} # }