verbatim_logical {yaml}	R Documentation

Alternalte logical handler

Description

A yaml handler function that will cause logical vectors to emit true/false value instead of yes/no value

Usage

verbatim_logical(x)

Arguments

x
logical values to convert to true or false.

Details

To use pass into as.yaml as part of a handler arguement list like list(logical=verbatim_logical).

Value

Returns a vector of strings of either true or false of class verbatim.

Author(s)

Charles Dupont and James Goldie (jimjam-slam)

See Also

as.yaml

Examples

vector <- c(TRUE, FALSE, TRUE)

as.yaml(vector, handlers=list(logical=verbatim_logical))

[Package yaml version 2.3.9 Index]