verbatim_logical {yaml}	R Documentation

Alternative logical handler

Description

A yaml handler function that causes logical vectors to emit true/false instead of yes/no values.

Usage

verbatim_logical(x)

Arguments

x
logical vector to convert to true/false.

Details

Pass this function to as.yaml() as part of the handler argument list like list(logical = verbatim_logical).

Value

Returns a vector of strings of either true or false of class verbatim.

Author(s)

Charles Dupont and James Goldie (jimjam-slam)

See Also

as.yaml

Examples

vector <- c(TRUE, FALSE, TRUE)

as.yaml(vector, handlers=list(logical=verbatim_logical))

[Package yaml version 2.3.10 Index]